fix reconet & some whitespace errors
Python reserves `lambda` as a keyword, so no function parameter can be
called `lambda`. However, the original workaround for this would just
replace argument names which contain `lambda` with `llambda`, even names
such as `no_shared_lambda` would become `llambda`. Which leads to a bug
when there are multiple arguments with names containing lambda, such as
in reconet.